+++ DataAddons/Retail/10_Dragonflight/AchievementData.lua
@@ -1,6 +1,4 @@
-local _, addon = ...
-local shared = addon.Data.AchievementData.Shared
-local Ach = shared.Ach
+local Ach = KrowiAF.Ach
 local faction = KrowiAF.Enum.Faction
 
 KrowiAF.AchievementData["10_00_00"] = {